Return the payload instead of the BLED112 packet
When using the read_attr and write_attr functions, only return the
payload (without the 4 byte L2CAP header and 1 byte payload length).
Previously, both functions returned the whole BLED112 packet, which
also contained the type, the class code and the command code.

Additionally, improve the version checking code.

--
This is part of the preparation to add a bluepy backend by adapting the
BLED112 interface to the bluepy.btle.Peripheral.