Move device scanning to the BLED112 related code
Create a scan function which returns the MAC address of a Myo armband by
specifying the corresponding 128 bit UUID of the control service and
optionally the MAC address to search for. This ensures that the provided
MAC address belongs to a Myo armband and that the device is online.
Additionally, adapt the signature of the connect function to match the
bluepy.btle.Peripheral.connect function.

--
This is part of the preparation to add a bluepy backend by adapting the
BLED112 interface to the bluepy.btle.Peripheral.