Remove BLED112 related code from handle_data
Change the handle_data() function arguments to expect a Bluetooth LE
attribute/cHandle and a payload instead of the former BLED112 packet.
Add a new wrapped_handle_data() function to ensure that these arguments
are extracted from a BLED112 packet and forwarded to the handle_data()
function. Furthermore, move code related to the processing of a BLED112
packet to the wrapped_handle_data() function.

--
This is part of the preparation to add a bluepy backend by adapting the
BLED112 interface to the bluepy.btle.Peripheral.