Add support for native Bluetooth interfaces
Add the option to connect the Myo armband using a standard Bluetooth
interface on Linux with bluepy instead of using the provided Bluegiga
BLED112 dongle.
Rename the former Bluetooth class to BLED112 and the instance name to
backend in order to differentiate the Bluetooth backends.
Add bluepy as an optional dependency to the setup.py file.